backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
httpx==0.27.2
google-generativeai==0.3.2
python-dotenv==1.0.1